Crea las claes que seran representadas como tablas en la base de datos, y hace las respectivas acciones en la clase principal para guardar los libros con sus datos

+++ src/main/java/com/project/LiterAlura/model/DatosLibro.java
@@ -10,7 +10,7 @@ public record DatosLibro(
         @JsonAlias("title") String titulo,
         @JsonAlias("authors") List<DatosAutor> autor,
         @JsonAlias("languages") List<String> idioma,
-        @JsonAlias("download_count") Integer descargas) {
+        @JsonAlias("download_count") Double descargas) {
 
     @Override
     public String toString() {
